+++ src/app/home/home.component.ts
@@ -16,7 +16,6 @@ export class HomeComponent implements OnInit {
   cards = [];
   slides: any = [[]];
   private someid = [];
-  thisshow = true;
 
   /**
    * format data to populate cards
@@ -32,7 +31,7 @@ export class HomeComponent implements OnInit {
   ngOnInit(): void {
     this.load_profiles();
     const parentThis = this;
-    this.someid = parentThis.userProfileService.getCurrentUser().id;
+    this.someid = parentThis.userProfileService.getCurrentUser().id;  // don't want to show the current users profile to themselves
     this.userProfileService.getAllProfiles().then((results) => {
       parentThis.load_cards(results, this.someid);
     });
@@ -46,9 +45,10 @@ export class HomeComponent implements OnInit {
   /**
    * Push data into the carosel cards
    */
-  load_cards(results, someid): void {
+  async load_cards(results, someid): Promise<void> {
     this.cards = [];
     let i;
+    let noShowid;
     for (i = 0; i < results.length; i++) {
       this.cards.push({buttonText: 'Expand Profile', });
       this.cards[i].title = results[i].get('name');
@@ -57,42 +57,27 @@ export class HomeComponent implements OnInit {
       this.cards[i].rank = results[i].get('rank');
       this.cards[i].username = results[i].get('username');
       this.cards[i].expand = false;
-      
-      if (typeof (results[i].get('user_id')) !== 'undefined' && this.liked2(results[i].get('user_id').id)){
-        console.log('here');
-        this.cards[i].show = false;
-      } else {
-        this.cards[i].show = true;
-      }
-      if (typeof (results[i].get('user_id')) !== 'undefined' && results[i].get('user_id').id === someid){
-        this.cards[i].show = false;
-      }
-      if(typeof (results[i].get('user_id')) !== 'undefined'){
-        console.log(this.liked2(results[i].get('user_id').id));
-        this.cards[i].user_id = results[i].get('user_id').id;
-      }
+      this.cards[i].show = true;
 
-      // if (typeof (results[i].get('user_id')) !== 'undefined'){
-      //   const parentThis = this;
-      //   const current_user_id = parentThis.userProfileService.getCurrentUser().id;
-      //   const thisid = results[i].get('user_id').id;
-      //   this.chatService.getAllChats().then((somes) => {
-      //     somes.forEach(element => {
-      //       let userA = element.get('userA');
-      //       let userB = element.get('userB');
-      //       let chatActivity = element.get('activeChat');
-      //       if (userA.id === thisid && userB.id === current_user_id && chatActivity === 'yes'){
-      //         this.thisshow = false;
-      //         console.log(thisid, this.thisshow);
-      //       } else if (userA.id === thisid && userB.id === current_user_id && chatActivity === 'invalid'){
-      //         this.thisshow = false;
-      //       } else {
-      //         console.log(thisid, this.thisshow);
-      //       }
-      //     });
-      //   });
-      //   this.cards[i].user_id = results[i].get('user_id').id;
-      // }
+      // noShowid will return true if there is a chatid returned indicating that the user has liked or disliked
+      // the current id (results[i].user_id)
+      if (typeof (results[i].get('user_id')) !== 'undefined'){
+        this.cards[i].user_id = results[i].get('user_id').id;
+        noShowid =  await this.chatService.getChatbyUser(results[i].get('user_id').id).then((moreresults) => {
+          if (moreresults.length === 1){
+            console.log(moreresults[0].id);
+            return true;
+          } else{
+            return false;
+          }
+        });
+        if (noShowid || results[i].get('user_id').id === someid){
+          this.cards[i].show = false;
+          console.log(results[i].get('user_id').id);
+          console.log(noShowid);
+        }
+
+        }
 
     }
   }
@@ -126,7 +111,11 @@ export class HomeComponent implements OnInit {
     this.slides = this.chunk(this.cards, 3);
   }
 
-    liked(likedid){
+  /**
+   * Profile is liked so chat goes from wait to yes which will indicate that this profile should not be
+   * shown.
+   */
+  liked(likedid){
     const parentThis = this;
     const current_user_id = parentThis.userProfileService.getCurrentUser().id;
     this.chatService.getAllChats().then((results) => {
@@ -138,38 +127,18 @@ export class HomeComponent implements OnInit {
             element.set('activeChat', 'yes');
             element.save().then((element) => {
               console.log('chat saved');
-              parentThis.log_data(current_user_id);
+              window.location.reload();
             });
-        } else if (userA.id === likedid && userB.id === current_user_id && chatActivity === 'yes'){
-              this.thisshow = false;
-        } else {
-              this.thisshow = true;
         }
       });
     });
   }
 
-  liked2(likedid){
-    const parentThis = this;
-    const current_user_id = parentThis.userProfileService.getCurrentUser().id;
-    this.chatService.getAllChats().then((results) => {
-      results.forEach(element => {
-        let userA = element.get('userA');
-        let userB = element.get('userB');
-        let chatActivity = element.get('activeChat');
-        if (userA.id === likedid && userB.id === current_user_id && chatActivity === 'wait'){
-            
-        } else if (userA.id === likedid && userB.id === current_user_id && chatActivity === 'yes'){
-            console.log('should return true', likedid);
-            return true;
-        } else {
-        }
-      });
-    });
-    return false;
-  }
-
-    disliked(dislikedid){
+  /**
+   * Profile is disliked so chat goes from wait to invalid which will indicate that this profile should not be
+   * shown.
+   */
+  disliked(dislikedid){
     const parentThis = this;
     const current_user_id = parentThis.userProfileService.getCurrentUser().id;
     this.chatService.getAllChats().then((results) => {
@@ -181,10 +150,8 @@ export class HomeComponent implements OnInit {
             element.set('activeChat', 'invalid');
             element.save().then((element) => {
               console.log('chat saved');
-              parentThis.log_data(current_user_id);
+              window.location.reload();
             });
-        } else if (userA.id === dislikedid && userB.id === current_user_id && chatActivity === 'invalid'){
-            return true;
         }
       });
     });